+++ crates/ui-components/src/lib.rs
@@ -81,6 +81,7 @@ pub mod labeled_value;
 pub mod link;
 pub mod link_button;
 pub mod listbox;
+pub mod listbox_item;
 pub mod logic_button;
 pub mod menu;
 pub mod menu_item;
@@ -220,6 +221,7 @@ pub use labeled_value::{LabeledValue, LabeledValueOrientation, LabeledValueTone}
 pub use link::Link;
 pub use link_button::LinkButton;
 pub use listbox::ListBox;
+pub use listbox_item::ListBoxItem;
 pub use logic_button::{LogicButton, LogicButtonVariant};
 pub use menu::Menu;
 pub use menu_item::MenuItem;